requirements.txt
langchain==0.0.184
openai==0.27.1
pandas==1.4.4
pdfminer.six==20221105
PyPDF2==3.0.1
streamlit==1.19.0
tiktoken==0.2.0
faiss-cpu
openpyxl